Finally fix bad values for int64u_t.  I accidentally had the wrong forced coversion so it was silent.

+++ intXu_t.h
@@ -191,7 +191,7 @@ class IntXu_t
           ss << "-";
         }
 
-        ss << (int) value;
+        ss << (int64_t) value;
 
         if ( uncertain )
         {
